Moves all photo's off the VIRB into a single folder
Renames the photo's from 123_VIRB\VIRB1234.JPG to
yyyy-mm-dd_123_VIRB1234.JPG so there is no naming conflict and sorting
is in chronological order.
After importing it's easy to apply the time_split.py on this folder to
split up into sequences.
Maybe this script works for other devices besides the VIRB, but I have
only tested it on the VIRB.
Uploading this script because it may be useful to other users.